Add naive PCA method
